+++ CHANGELOG.md
@@ -19,7 +19,7 @@
   (https://codeberg.org/dnkl/yambar/issues/77).
 * bar: `layer: top|bottom`, allowing the layer which the bar is
   rendered on to be changed. Wayland only - ignored on X11.
-* river: `per-output: false|true`.
+* river: `all-monitors: false|true`.
 * `-d,--log-level=info|warning|error|none` command line option
   (https://codeberg.org/dnkl/yambar/issues/84).
 * river: support for the river-status protocol, version 2 (‘urgent’